Fix implicit cast to float32

+++ muvi/tools/utils.py
@@ -610,7 +610,9 @@ def to_mdata(
 
     adata_dict = {}
     for vn in model.view_names:
-        adata = ad.AnnData(obs_dict[vn], varm={loadings_key: loadings_dict[vn].T})
+        adata = ad.AnnData(
+            obs_dict[vn], varm={loadings_key: loadings_dict[vn].T}, dtype=np.float32
+        )
         if betas_dict:
             adata.varm[betas_key] = betas_dict[vn].T
         if masks_dict:
